doc-extract-starter/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
pdfminer.six==20231228
PyYAML==6.0.1
regex==2024.5.15
rapidfuzz==3.9.6
